% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_ped_path}
\alias{example_ped_path}
\title{Get Path to Example Pedigree CSV File}
\usage{
example_ped_path()
}
\value{
Character scalar. Full path to \code{family_pedigree.csv}.
}
\description{
Returns the path to the bundled example pedigree CSV file for the FAM_KHAN
family.
}
\examples{
ped_path <- example_ped_path()
\dontrun{
ped <- read_pedigree(ped_path)
}

}
\seealso{
\code{\link{example_vcf_dir}}, \code{\link{read_pedigree}}
}
